+++ tests/NW.UnivariateForecasting.UnitTests/Forecasts/ForecastingInitTests.cs
@@ -76,7 +76,7 @@ namespace NW.UnivariateForecasting.UnitTests.Forecasts
 
         [TestCaseSource(nameof(forecastingInitTestCases))]
         public void ForecastingInit_ShouldCreateAnInstanceOfThisType_WhenProperArgument
-            (string observationName, List<decimal> values, decimal? coefficient, decimal? error)
+            (string observationName, List<double> values, double? coefficient, double? error)
         {
 
             // Arrange
@@ -95,13 +95,13 @@ namespace NW.UnivariateForecasting.UnitTests.Forecasts
             if(actual.ObservationName != null) 
                 Assert.IsInstanceOf<string>(actual.ObservationName);
 
-            Assert.IsInstanceOf<List<decimal>>(actual.Values);
+            Assert.IsInstanceOf<List<double>>(actual.Values);
 
             if (actual.Coefficient != null)
-                Assert.IsInstanceOf<decimal?>(actual.Coefficient);
+                Assert.IsInstanceOf<double?>(actual.Coefficient);
 
             if (actual.Error != null)
-                Assert.IsInstanceOf<decimal?>(actual.Error);
+                Assert.IsInstanceOf<double?>(actual.Error);
 
         }
